requirements.txt
anthropic==0.40.0
pynput==1.7.6
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
python-dotenv==1.0.0
